realtek: switch target to upstream platform
Use the upstream MACH_REALTEK_RTL platform instead of the downstream
RTL83XX platform. This effectively deprecates the original platform.

The updated config was generated by kernel_oldconfig, using the defaults
for the now available symbols.

Since the upstream platform is based on MIPS_GENERIC, the kernel load
and entry addresses need to be updated to match the expected values.

Co-developed-by: INAGAKI Hiroshi <musashino.open@gmail.com>
Signed-off-by: INAGAKI Hiroshi <musashino.open@gmail.com>
Signed-off-by: Sander Vanheule <sander@svanheule.net>
Tested-by: Bjørn Mork <bjorn@mork.no>