realtek: initialise mulithreading on supported SoCs
In addition to CPS SMP setups, also try to initialise MT SMP setups with
multiple VPEs per CPU core on MIPS_GENERIC. CMP SMP support is not
provided as it is considered deprecated. This is required to enabled SMP
on RTL839x and RTL930x devices.

Additionally, rework the code by dropping the err variable and make it
similar to how other platforms perform this initialisation.

Co-developed-by: INAGAKI Hiroshi <musashino.open@gmail.com>
Signed-off-by: INAGAKI Hiroshi <musashino.open@gmail.com>
Signed-off-by: Sander Vanheule <sander@svanheule.net>