realtek: Switch DGS-1210-10P DTS to gpio.dtsi
The DTS file for the DGS-1210-10P is slightly different from the other
DGS-1210 devices, in that it didn't specify a gpio-restart node when it
was added. Add a label to the gpio-restart node so we can reference and
disable it from the DGS-1210-10P's DTS file, so maintain the current
reboot behaviour.

This converts the last device from the product family to the common
definition for the (external) GPIOs.

Signed-off-by: Sander Vanheule <sander@svanheule.net>