realtek: Switch ApresiaLightGS120GT-SS RTL8231 driver
Switch the implementation for the RTL8231 GPIO expander to the new
driver.

This allows specifying the GPIO driving the RTL8231's reset as a proper
MDIO reset line, so the gpio-hog can be dropped. Since it was pinned at
a high level, the reset line is actually active-low (i.e. high when not
in reset).

Since the GPIO used for the reset key is capable of hardware debouncing,
also add the (only supported) debounce interval of 100ms.

Signed-off-by: Sander Vanheule <sander@svanheule.net>